requirements.txt
numpy
matplotlib
openml-python
git+https://github.com/openml/openml-python-contrib.git
pandas
scipy
scikit-learn
